+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Add node versions `17` and `18` to the Github workflow
+* Add animation instead of `no-message` icon in `chat.js` - [ripe-robin-revamp/#380](https://github.com/ripe-tech/ripe-robin-revamp/issues/380)
 
 ### Fixed